refactor: separate HTTPClient behaviour from Req implementation
